Remove Dependency on Newtonsoft.Json From Protocol objects
When trimming or native AOT'ing an app, we want to remove Newtonsoft.Json dependencies when possible because it brings in a lot of AOT warnings and adds a lot of size to the final app.

Begin by removing these simple cases in the ToStrings of JsonRpcError, Request, and Result. ToString methods of instantiated types can never be trimmed, so the code in them is always preserved.